Add SAP and real basis output

+++ output/makeinp_sap_gto.sh
@@ -3,11 +3,15 @@
 . ../mets.sh
 
 atom=$(basename $PWD)
-cat > atoms.xyz <<EOF
+if [[ -f ../${atom}_0.occs ]]; then
+    if [[ ! -f atoms.xyz ]]; then
+        cat > atoms.xyz <<EOF
 1
 $atom
 $atom 0 0 0
 EOF
+    fi
+fi
 
 p=${PWD}
 for((iB=0;iB<${#Bvals_gto[@]};iB++)); do
@@ -16,20 +20,22 @@ for((iB=0;iB<${#Bvals_gto[@]};iB++)); do
 	mkdir -p $wdir
     fi
     cd $wdir
-    for((iocc=8;iocc>=0;iocc--)); do
+    for((iocc=0;;iocc++)); do
 	   occfile="../../${atom}_${iocc}.occs"
-           if [[ -f ${occfile} ]]; then
+           if [[ ! -f ${occfile} ]]; then
+               break
+           fi
 
-               # Find out multiplicity
-               M=$(cat ${occfile} |awk 'BEGIN {na=0; nb=0} {na+=$1; nb+=$2} END {print na-nb+1}')
-               mmax=$(cat ${occfile} |awk 'BEGIN {mmax=0} {if($3>mmax) {mmax=$3}} END {print mmax}')
-	       
-               # GTO calculations
-               for basis in ${bases[@]}; do
-		   cat > ${basis}_${iocc}.run <<EOF
-System ../atoms.xyz    
-Basis ${basis}	       
-Decontract *	       
+           # Find out multiplicity
+           M=$(cat ${occfile} |awk 'BEGIN {na=0; nb=0} {na+=$1; nb+=$2} END {print na-nb+1}')
+           mmax=$(cat ${occfile} |awk 'BEGIN {mmax=0} {if($3>mmax) {mmax=$3}} END {print mmax}')
+
+           # GTO calculations
+           for basis in ${bases[@]}; do
+               cat > sap_${basis}_${iocc}.run <<EOF
+System ../atoms.xyz
+Basis ${basis}
+Decontract *
 Method HF
 LinearOccupations -1
 LinearOccupationFile ${occfile}
@@ -41,12 +47,11 @@ OptLM false
 Complexbas true
 MaxInitIter 100
 EOF
-		   conv=$(grep "Converged to" sap_${basis}_${iocc}.stdout)
-		   if [ -z "$conv" ]; then
-                       echo ${atom} ${basis} ${iocc} ${Bvals_gto[iB]}
-                       erkale_complex_orbs_omp ${basis}_${iocc}.run &> sap_${basis}_${iocc}.stdout
-		   fi
-               done
-	   fi
+	       conv=$(grep "Converged to" sap_${basis}_${iocc}.stdout)
+               if [ -z "$conv" ]; then
+                   echo ${atom} ${basis} ${iocc}
+                   erkale_complex_orbs_omp sap_${basis}_${iocc}.run &> sap_${basis}_${iocc}.stdout
+               fi
+           done
        done
        done
